backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
pydantic>=2.9
numpy>=1.26
opencv-contrib-python>=4.10
scipy>=1.13
ruamel.yaml>=0.18
pyinstaller>=6.11
